feat(chao): Add code to create a Regular Two-tone Chao Egg
This commit adds code to create a Regular Two-tone Chao Egg with random statistics and store it in the database. It also sends a message to the user to notify them about the newly received Chao Egg and waits for 5 seconds before hatching it into a Chao. The hatched Chao is then stored in the database with updated information and another message is sent to the user to notify them about the hatching.